backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-multipart==0.0.20
httpx==0.28.1
chromadb==0.6.3
watchdog==6.0.0
PyMuPDF==1.25.5
python-docx==1.1.2
openpyxl==3.1.5
Pillow==11.2.1
tiktoken==0.9.0
